+++ src/lib/routes.js
@@ -1,5 +1,6 @@
 const debug = require('debug')('ExpressImpRouter.routes');
 const routeExtractor = require('../extract/routes');
+const clone = require('lodash.clone');
 
 let isDebug = false;
 let routes = [];
@@ -14,60 +15,63 @@ function get(debug) {
   debug = debug || false;
 
   if(!debug) {
-    return routes.filter(route => route.generated);
-  } else {
-    const colors = require('colors');
-    colors.setTheme({
-      warn: 'yellow',
-      error: 'red'
-    });
-    let color = colors.white, status = '';
-
-    const debugRoutes = [];
-    let currentRoutes = [];
-    let currentConfigName = null;
-
-    routes
-    .sort((a, b) => {
-      const textA = a.config_name.toUpperCase();
-      const textB = b.config_name.toUpperCase();
-      return (textA < textB) ? -1 : (textA > textB) ? 1 : 0;
-    })
-    .map((obj) => {
-      if(!obj.generated) {
-        color = colors.red;
-        status = 'Route not generated';
-      } else {
-        color = colors.green;
-        status = 'Route generated :';
-        if(obj.debug.level) {
-          color = colors[obj.debug.level];
-        }
-      }
-
-      if(!currentConfigName) {
-        currentConfigName = obj.config_name;
-      }
-
-      if(currentConfigName && currentConfigName !== obj.config_name) {
-        debugRoutes.push(currentRoutes);
-        currentRoutes = [];
-        currentConfigName = obj.config_name;
-      }
-
-      currentRoutes.push({
-        status: color(status),
-        method: color(obj.method.toUpperCase()),
-        route: color(obj.route),
-        controller: color(obj.debug.controller ? obj.debug.controller : 'N/A'),
-        action: color(obj.debug.action ? obj.debug.action : 'N/A'),
-        message: color(obj.debug.message ? obj.debug.message : 'N/A'),
-      });
-    });
-
-    debugRoutes.push(currentRoutes);
-    return debugRoutes;
+    return routes.map((route) => {
+      route.debug.controllersNotGenerated = route.controllers.filter(controller => !controller.generated);
+      route.actions = route.controllers.filter(controller => controller.generated).map(controller => controller.action);
+      return route;
+    }).filter(item => item.actions && item.actions.length > 0);
   }
+
+  const colors = require('colors');
+  colors.setTheme({
+    warn: 'yellow',
+    error: 'red',
+    success: 'green',
+  });
+
+  return [routes.map((route) => {
+    let color = colors.white;
+    const controllersName = [], actionsName = [], debugMessages = [];
+    let controllersGenerated = 0, status = '';
+
+    route.controllers.map((controller) => {
+      if(controllersName.indexOf(controller.controllerName) === -1) {
+        controllersName.push(controller.controllerName);
+      }
+
+      if(actionsName.indexOf(controller.actionName) === -1) {
+        actionsName.push(controller.actionName);
+      }
+
+      if(controller.debug.message) {
+        debugMessages.push(controller.debug.message);
+      }
+
+      if(controller.generated) {
+        controllersGenerated++;
+      }
+    });
+
+    if(controllersGenerated === 0) {
+      color = colors.error;
+      status = 'Nothing generated';
+    } else if(controllersGenerated !== route.controllers.length) {
+      color = colors.warn;
+      status = 'Partial generated';
+    } else {
+      color = colors.success;
+      status = 'Generated';
+    }
+
+    return {
+      status: color(status),
+      method: color(route.method.toUpperCase()),
+      route: color(route.route),
+      controller: color(controllersName),
+      action: color(actionsName),
+      message: color(debugMessages),
+    };
+  })];
 }
 
 /**
@@ -80,10 +84,6 @@ function extract(configAll) {
   for(const i in configAll) {
     let scopeRoutes = [];
     scopeRoutes = scopeRoutes.concat(routeExtractor.route(configAll[i], configAll[i].routes, isDebug));
-
-    if(configAll[i].config && configAll[i].config.route_mode === 'strict') {
-      scopeRoutes = dedupe(scopeRoutes);
-    }
 
     for(const j in scopeRoutes) {
       scopeRoutes[j].config_name = `Config${i}`;
@@ -100,38 +100,4 @@ function extract(configAll) {
  */
 function enableDebug() {
   isDebug = true;
-}
-
-/**
- * Dedupe routes
- */
-function dedupe(scopeRoutes) {
-  debug('Start deduping');
-  const routeValidated = [];
-  let index = -1;
-
-  for(const i in scopeRoutes) {
-    if(scopeRoutes[i].static) {
-      continue;
-    }
-    const strToCompare = scopeRoutes[i].method+scopeRoutes[i].route;
-    const strToCompareHard = scopeRoutes[i].method+scopeRoutes[i].route.replace(/(\(.+\))/gi, '');
-
-    if((index = routeValidated.indexOf(strToCompare)) !== -1) {
-      scopeRoutes[i].debug.message = `This route already exists : ${scopeRoutes[i].method.toUpperCase()} ${scopeRoutes[i].route}`;
-      scopeRoutes[i].debug.level = 'error';
-      scopeRoutes[i].isDedupe = true;
-    } else {
-      if((index = routeValidated.indexOf(strToCompareHard)) !== -1) {
-        scopeRoutes[i].debug.message = `Maybe there is similar route : ${scopeRoutes[i].method.toUpperCase()} ${scopeRoutes[i].route} :: ${routeValidated[index].replace(/([a-z]+)(\/.+)/, '$1').toUpperCase()} ${routeValidated[index].replace(/([a-z]+)(\/.+)/, '$2')}`;
-        scopeRoutes[i].debug.level = 'warn';
-        scopeRoutes[i].generated = true;
-      }
-
-      scopeRoutes[i].isDedupe = false;
-      routeValidated.push(scopeRoutes[i].method+scopeRoutes[i].route);
-    }
-  }
-  debug('Deduping : done');
-  return scopeRoutes;
 }