+++ CHANGELOG.md
@@ -39,6 +39,9 @@
 
 * View engine handling
 
+* Debug
+  * when multiple actions/controllers in one route
+
 * Tests
 
 * Docs